requirements/base.txt
Django==2.0.4
# API building
djangorestframework==3.8.2
Pillow==5.1.0
redis==2.10.6

# Running workerized django
gunicorn==19.7.1

# Progress-bar for books downloading progress
tqdm==4.19.5

# Own libraries
moonreader_tools==1.1.0
PyYAML==3.12

psycopg2==2.7.4

# Markdown editor in the admin page
django-pagedown==1.0.4

django-environ==0.4.4

# Temporary workaround, as current version does not support Django 2.0
# and we're cool guys that use latest lib versions
-e git://github.com/mrlokans/django-admin-honeypot.git@2241626fe6c678396ae1ff53dd9122a59d173a6c#egg=django_admin_honeypot

django-health-check==3.5.1

# Exporting data from the admin
django-import-export==1.0.0

# Manage.py extra commands
django-extensions==2.0.6

# Reading env variables
django-dotenv==1.4.2

# telegram integration
python-telegram-bot==10.1.0

-e git://github.com/MrLokans/onliner-agent-finder.git@0.4.1#egg=agent_spider

# Blogging support
django-fluent-blogs==2.0.2
django-fluent-contents[text,code,markdown]==2.0.3
docutils==0.14
Markdown==2.6.11
textile==3.0.3
django-fluent-utils==2.0
django-wysiwyg==0.8.0
django-taggit==0.22.2